Fix signed/unsigned comparisons
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@14258 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/classes/stack.h
@@ -99,7 +99,7 @@ namespace Firebird {
 			{
 				fb_assert(value >= 0);
 				
-				if (value <= inherited::getCount())
+				if ((size_t) value <= inherited::getCount())
 					return true;
 
 				for(const Entry* stk = this; stk && value > 0; stk = stk->next)
@@ -190,7 +190,7 @@ namespace Firebird {
 
 				if (elem)
 				{
-					if (value < elem)
+					if ((size_t) value < elem)
 						return true;
 				
 					value -= elem - 1;
